The final commit
I ran out of credits and have no clue if the backend works. Only God will know what will happen when this is up on Render. 🙏

1. App.jsx (Frontend)
a. Added a lot of const variables to handle the microphone feature.
- minutes: audio transcription
- minutesLoading: check if file is bring transcribed, disables button if True
- voiceMode: check if voice mode is turned on
- isSpeaking: check if chatbot is speaking in text to speech
- isListening: True when the microphone is actively listening for speech input
-micCooldown: True when the microphone is in a cooldown period to prevent spamming
-recognitionRef: Reference to the current speech recognition instance (for controlling the mic)

b. Changed/added conditional keywords
- conditionalRelated: more specific
- sightRelated: added for visually impaired users

c. Added backend checks so both Frontend and Backend retain the user's special needs condition (until the page is reloaded)

d. Added improvements to the microphone
- Cooldown to prevent users from mashing the button and causing bugs
- Microphone + voice mode activates automatically upon user saying that they are visually impaired
- Added feedback to the microphone so that the user knows when it is listening and when it is not
- Microphone can be activated using CTRL + Spacebar keys
- When microphone is active, the textbox temporarily disappears

2. app.py (Backend)
a. Added the accessibility features to the chatbot so the chatbot can tell the user more about them

b. Added conditionals to update system prompt for all 3 disabilities covered.

c. Added backend function to detect what special needs condition the user has (possibly incomplete but I think it should help if the Frontend is unable to do so)

d. Added Text to Speech function that triggers when voice mode in the front end is activated. Voice is fixed to "alloy" but can be changed

3. App.css (CSS for Frontend)
- Added CSS for the button that activates text to speech (Voice Mode) for the chatbot.
- Placed it on the bottom right corner of the application so that the user can access it easily (without clicking on it accidentally)